backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlmodel==0.0.16
sqlalchemy==2.0.30
pydantic>=2.8,<3
python-dotenv==1.0.1

# AI Providers
openai==1.14.0
anthropic==0.25.4
deepseek==1.0.0
mistralai==1.0.2
groq>=0.9
